Semi-automate transport queue cleanup operation
Summary:
There is already a P0 committed H2 OKR our team is working on to improve operations due to such slowdowns, but that change could be invasive as it touches more internal parts of core Cassandra, and we do not want to rush that change. Nobody in the industry has tried out such a change before. So, the attempt in this ticket is to do a minimal change, which is safe and provides some relief in operations and/or mitigation time!

Please note, this is not prevention; it only helps make operations easy so that faster mitigation.

Architecturally, Cassandra's transport queue is FIFO. But when it is bombarded extremely hard by the callers, either due to expensive queries, then the queue grows, and then in the worst case scenario, we would have to restart to drain the queue, which increases the mitigation and adds enormous complexity in handling on-call operations.
So, the thing we want to do in this task is as follows:
There is no way we can avoid expensive queries, etc., from the callers, so this task does not do any sort of prevention action. That needs to be addressed by the callers, etc. What we can do is automate our operations, which will reduce the mitigation time after the issue happens.
We want to do two architectural changes to the OSS Cassandra's behavior:
1. Add a new timeout so that we can discard the messages from the queue that timeout from the queue
2. An automated workflow that our team can run to clean up the queue without the restart.

Both of the above things are pretty new as this has not been tried out before in the Cassandra community at scale, so we have to be extremely careful with the rollout, etc., to ensure it does not introduce any new behavior.

- An existing nodetool workflow should work out of the box for #2 above, so it is not covered as part of this task. It will be validated separately.

Test Plan:
- A detailed unit+integration test case was added to the diff that captures the end-to-end behavior.
- Test case framework for Cassandra CLI (Nodetool) are not present in the OSS as it does not fall in the hot path, it is only for CLI purposes, such as debugging, etc.

Reviewers: #ldap_cassandra-super-users, O2017 Cassandra Service, yuqiy

Reviewed By: #ldap_cassandra-super-users, O2017 Cassandra Service, yuqiy

Subscribers: gagarwal, yuqiy

JIRA Issues: SO-32855

Differential Revision: https://code.uberinternal.com/D14511861