Reduce timeouts in upgrade_through_versions_test due to ccm downloads/building
In docker we want to use the existing populated ~/.ccm/repositories but don't want to write to the container's filesystem (it can leak fd).  So we set CCM_CONFIG_DIR to be in the run-python-dtest.XXXXXX tmpdir.  Previously this was done in the base (non-docker) script, but this is only needed in docker (and outside of docker we want the normal ~/.ccm usage)

 patch by Mick Semb Wever; reviewed by Brandon Williams for CASSANDRA-19813