Remove some coded exceptions out of OtherExceptions
Summary: Before we added metrics for each coded exception, we use OtherException to capture all exceptions we are not interested in. Now we have a metric for each coded exception, OtherException is kind of confusing. With this diff, OtherException only captures exceptions which are not coded and so will be not expected: if it shows up, there must be wrong in the code.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Subscribers: yuqiy

JIRA Issues: SO-33082

Differential Revision: https://code.uberinternal.com/D14573019