Shuffle keyspaces and tables when running auto-repair
Summary: Given that we run auto-repair with parallelism greater than 1, it is likely that we may be running repair for the same keyspace/table on multiple nodes at the same time. In the worst case this may overload the online traffic for these tables. The diff updated the auto-repair scheduler to randomize the order of keyspaces/tables in repair operations to reduce the likelihood that different nodes run repair on the same table at the same time.

Test Plan:
Unit tests + Deployed the changes to a test instance and verified that different nodes use a different order for repair operations.

Node 1:
{F951012841}

Node 2:
{F951013151}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, jchovati

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, jchovati

Subscribers: jchovati

JIRA Issues: SO-32238

Differential Revision: https://code.uberinternal.com/D14626483