Prevent node restart from scheduling a premature auto-repair cycle
Summary:
When the auto-repair scheduler determines whether or not it should start an auto-repair cycle, it refers to the timestamp showing when the current node was repaired last. If this timestamp is older than a given duration (e.g. 24 hours), the scheduler will start auto-repair.

However, with the current implementation of the scheduler, this timestamp is taken only from memory, we do not take into account the node's repair history within the `system_auto_repair` tables. This means that when a given node is restarted, the last repair time kept in memory is 0 even if the node had been repaired recently. As a result, restarting the node can cause a premature auto-repair cycle which can lead to 2 issues:
  # A Cassandra restart should be a very safe operation. We frequently rely on node restarts as a mitigation strategy during incidents. Having the node start auto-repair immediately after it boots exposes the node to additional risk while it may already be in a compromised state.
  # Restarting a node results in the node being unavailable for a short period of time. During this time the node is not accepting writes and becomes desynchronized with the rest of the ring. While running anti-entropy repair can resolve this kind of desynchronization, a much more efficient mechanism in such cases is hinted handoff. Therefore we should rely on hinted handoff rather than anti-entropy repair to resolve restart-related desynchronization.

This diff improves the auto-repair scheduler to retrieve the timestamp of the last auto-repair cycle for the node from the system tables after the node boots. As a result, a node restart no longer has the capacity to trigger a premature auto-repair cycle.

Test Plan:
Unit tests + deployed the changes to a test instance:

**Before:**
{F953945585}

**After:**
{F953954803}

**Scheduler is still able to schedule auto-repair without a restart:**
{F954248697}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

JIRA Issues: SO-33471

Differential Revision: https://code.uberinternal.com/D14725619