Extend nodetool sstablerepairedset to affect all keyspaces when no keyspace is provided in args
Summary: Improves the nodetool command `sstablerepairedset` to allow modifying the SSTables of all tables on the node if no keyspace is provided when calling the command. This will allow migrating the entire dataset of a given node between repaired and unrepaired states with a single command.

Test Plan:
Deployed the changes to a test instance and tested:

  # `nodetool sstablerepairedset --is-unrepaired --really-set`
  # `nodetool sstablerepairedset --is-repaired --really-set keyspace1`
  # `nodetool sstablerepairedset --is-repaired --really-set`

Results are shown below:

{F959417759}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, dalin

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, dalin

Subscribers: dalin

JIRA Issues: SO-33559

Differential Revision: https://code.uberinternal.com/D14835515