Backport the sstablepartitions offline tool from 5.0 to find large partitions in sstables
Summary: This diff contains the backporting of sstablepartitions from open source as is. It does not have any custom improvements from us yet (such as extra error handling to handle crashes due to compaction).

Test Plan:
Running on a file, human-readable:
{F974399707}
Running on a file, machine-readable:
{F974394525}
Running on a directory, machine-readable:
{F974400653}
Running on a directory, size threshold = 100KiB:
{F974956987}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Subscribers: yuqiy

JIRA Issues: SO-34666

Differential Revision: https://code.uberinternal.com/D15001873