Include jvm dtests in sonarqube code coverage
Summary:
code coverage in sonarqube is low because jvm dtests
are not included. In this diff, we include jvm dtests
which increases the coverage from 76% to 81%. We also
increase the timeout for the build as jvm dtests takes
additional 4 hours.

Below screenshot from sonarqube staging shows new coverage as 81%

{F969776661}

Staging build link : https://buildkite.com/uber/generic-udj/builds/3850916

Note that coverage shown in diffs (like this one) will not show improvement because
the diff builds uses separate scripts which are not modified to avoid
increasing build time. These builds (for diffs) execute all tests cases except that
the coverage for jvm dtests is not measured.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Subscribers: yuqiy, runtian

JIRA Issues: SO-34888

Differential Revision: https://code.uberinternal.com/D15067989