CNDB-10945: Change calculation of sstable span for small sstables (#1299)
In addition to correcting for very small spans, this also corrects sstable
spans for ones having a small number of partitions where keys can easily
fall in a small range. For these cases we use a value derived from the
number of partitions in the table, which should work well for all
scenarios, including wide-partition tables with a very limited number of
partitions.

 (Rebase of commit 921505a35622e46ffa0e22a76e95b5e636c8a0c2)

Adapted from datastax/cassandra commit 5a6f917bc3 for CASSANDRA-21615.
The estimated partition count is read from TableMetrics in
ShardManager.create because this tree has no CompactionRealm. Hunks for
ShardManagerReplicaAware and other files that do not exist here are
dropped; their logic moves to ShardManagerNoDisks constructor arguments.