Add observability in nodetool clearsnapshotfiles
Summary:
After the backup container started to use the `nodetool clearsnapshotfiles`, we observed that the call sometimes timed out or returned an error saying "file not found".

This obviously caused issue on backup container side. However, we need observability on C* side to understand the impact of this API call, especially when it times out, on C* servers. We need these metrics emitted on C* server side:

1. latency of the API call
2. success rate

Test Plan:
**end-2-end test:**
verified the value of the metrics via Jconsole

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian, kanoongo

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

JIRA Issues: SO-35054

Differential Revision: https://code.uberinternal.com/D15181857