Add error handling to SSTablePartitions to recover from sstable file deletion
Summary: Open source's sstablepartitions tool will crashed if it tries to scan an SSTable that is already deleted (the deletion can be due to an in-progress compaction). This is because the way the tool is implemented is quite simple: first, it fetches all directories and files from the input, and then it goes through those files one by one. So if an SSTables is removed after it was fetched but before it is read, the tool will crash when it attempts to read the sstable. More error handlings was added in this diff so that the tool won't crash when it attempts to read an SSTable that no longer exists. It will just simply skip to the next SSTable on its list.

Test Plan:
Testing plan:
  # Run cassandra-stress
  # Do nodetool flush thrice so that there are 3 SSTables
  # Inject sleep into various places of sstablepartitions' code (where the tool is scanning the sstable files)
  # Run sstablepartitions
  # While sstablepartitions is sleeping, run “nodetool compact ----user-defined <sstable1> <sstable2>” which merges 2 of the 3 SSTables into a single sstable
{F975504539}
As shown in the screenshot, the tool recovers gracefully from the error and moves on to other files.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Subscribers: yuqiy

JIRA Issues: SO-34666

Differential Revision: https://code.uberinternal.com/D15204569