Reduce log noise coming from repair
Summary: With the rollout of IR to our fleet, repair operations have created a very significant amount of new log messages which are making it difficult to track down other issues within the logs. This diff demotes some log messages which based on my experience produce a disproportionately large amount of noise and are not too useful for tracking down issues with repair.

Test Plan:
Deployed changes to a test instance and verified that:
  # We are still able to run and complete repair
  # Repair still has a 100% success ratio
{F979141075}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, gagarwal, yafeng

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, gagarwal, yafeng

Subscribers: yafeng, gagarwal

JIRA Issues: SO-35112

Differential Revision: https://code.uberinternal.com/D15280179