Implement streaming count-down latch for committing SSTables
Summary:
Introduces a count-down latch for streaming session which blocks committing individual stream tasks until all stream tasks within a given stream session are ready to commit. The reason why this is required is explored [[ https://docs.google.com/document/d/1n4KsFKRFl8S6t_6RjD6x7kxwP1_q4lPDcPj3Poi2U7o/edit?tab=t.0#heading=h.9c45gljcugr6 | here ]].

By default the latch has a timeout of 0s which means it will essentially be a NO-OP for now. This has been confirmed in unit tests and also manual tests described below. The plan is to collect some data to validate that the latch is working as expected at scale and also to determine a reasonable value for the timeout before setting the default timeout to a non-zero value.

Test Plan:
Unit tests + deployed changes to a staging instance and issued replacement.

The commit latch timed out (this is expected as the timeout is set to 0 by default):
{F991079923}

The maximum delay created by the latch was 1 ms when using a 0s timeout:
{F991079423}

**Test with long timeout:**

  # Set commit latch timeout to 1000s: https://odin.uberinternal.com/cadence/execution/odin-dca1/f34aec41-4031-47ce-9f73-cace8f7bc9b5/ce06830f-069b-4075-b193-30e82b573a64
  # Scheduled replacement: https://odin.uberinternal.com/cadence/execution/odin-dca1/1e071d3f-b084-4fa9-922c-8c57c0a351f4/3b426b4c-0b90-4747-a22d-44cda8e074e4

The joining node reported a max commit latch wait time of ~25ms:
{F995058019}
And no timeouts:
{F995058187}

2nd test:
  # Set commit latch timeout to 1000s: https://odin.uberinternal.com/cadence/execution/odin-dca1/e0cb7ccf-7c51-45d8-b0b2-38d07ea62837/5092e838-1952-4119-90d1-a902b23ce5c5
  # Scheduled replacement: https://odin.uberinternal.com/cadence/execution/odin-dca1/1019f80c-202c-4343-9536-815b489ec672/6e723b7f-ca45-413e-9030-e843df207099
  # Enabled debug logs for the joining node

Node printed out logs saying that it waited ~3 mins for the commit latch to be released:
{F995907445}

**Test for node crashing during bootstrapping:**
  * Set commit latch timeout to 1000s: https://odin.uberinternal.com/cadence/execution/odin-dca1/e0cb7ccf-7c51-45d8-b0b2-38d07ea62837/5092e838-1952-4119-90d1-a902b23ce5c5
  * Scheduled replacement: https://odin.uberinternal.com/cadence/execution/odin-dca1/0c9ff13b-b91d-4d00-b67c-a94cc5212561_7/7b4eb6b6-0513-45a7-a10d-b4e7d45b3d22
  * Enabled debug logs for the joining node
  * Waited for the log message saying that the node started waiting for the commit latch to be released:
{F996004253}
  * Identified the table id  in the logs above as the id for system_auth.roles:
{F996005033}
  * Found the transaction log for this streaming task and verified it does not contain a COMMIT entry:
{F996005257}
  * Also note that the table has only 1 SSTable, this means that the transaction log must be for the stream task that is waiting for the commit latch to be lifted. If it were for a different stream task, we would see at least 2 SSTables.
  * Killed the joining node before the commit latch was lifted:
{F996006075}
  * Once Cassandra rebooted verified that the node's logs contain messages about the system_auth.roles SSTable being deleted due to an unfinished transaction log:
{F996006509}
  * Verified that the SSTable was deleted after Cassandra rebooted (note that while after the restart we see an SSTable with the same name as the SSTable we had before the restart, the new SSTable has a later timestamp meaning that the old SSTable got deleted and Cassandra simply created a new SSTable with the same name):
{F996007895}
  * Also, metrics are showing that the SSTable was not added to the live set before or after the restart (which is what would happen without the commit latch):
{F996012007}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, gagarwal, yuqiy, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, gagarwal, yuqiy, runtian

Subscribers: runtian, gagarwal, yuqiy

JIRA Issues: SO-36789

Differential Revision: https://code.uberinternal.com/D15525595