Fix memory leak with BTree.FastBuilder
Summary:
There is a memory leak in BTree.FastBuilder.    This leak may only be seen in large clusters as originally reported.  The issue observed is the amount of heap memory used continues to grow until the node struggles with multiple full gc pauses.  The underlying issue is with BTree.FastBuilder.reset() method in that the object's buffer is not completely cleared in some scenarios.

This patch ensures that the reset() method fully clears the internal buffer to avoid leaving unused references in it.

Fix BTree.FastBuilder.reset() and test
Patch by benedict; reviewed by blambov for CASSANDRA-19785

Test Plan:
Using a 6 node cluster in prod, an attempt was made to reproduce the issue using the cassandra test tool to drive write/read traffic to the test cluster, observing the increase in memory usage, and then validating that the memory usage does not drop.  This was done without this patch applied and after this patch was applied.  Heap dumps were also taken before / during / after the stress test, but the behavior of expected in FastThreadLocalThread was not observed, potentially due to the fact that only a small cluster was used for the test.

There were no negative side effects observed during the test with the patched cluster nodes.

Results of the actions/observations for the tests are here https://docs.google.com/document/d/1qR8UUw4WfV6oBHdt_gQYvUUhvQ5qm9EHm8UEx6raFlM/edit?tab=t.0#heading=h.4qayofh53x8g

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

JIRA Issues: SO-36964, SO-38674

Differential Revision: https://code.uberinternal.com/D16033711