Fix StorageProxyWriteMutationSizeHistogram metric name
Summary:
It seems like the `MutationSizeHistogram` metric for writes was not following the naming convention for `StorageProxy` metrics. It was named  simply `MutationSizeHistogram` whereas other write metrics follow the naming convention `StorageProxyWrite<METRIC_NAME>`. In addition to that, we already have a `MutationSizeHistogram` for CAS writes named `StorageProxyCASWriteMutationSizeHistogram`.

This diff renames the `MutationSizeHistogram` metric for writes to use the standard naming convention. Note that this discrepancy in naming is not present in OSS Cassandra. The metric name is built using Uber-specific code.

Test Plan:
**jconsole results before:**

{F1017059175}

**jconsole results after:**
{F1017058581}

After deploying together with changes in D16085039 and D16085877:
{F1017068351}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, yuqiy

Subscribers: yuqiy

JIRA Issues: SO-34807

Differential Revision: https://code.uberinternal.com/D16085613