Fix race condition in AutoRepairState
Summary: Fixes a race condition in the `AutoRepairState` class where an already complete repair job may trigger a subsequent repair job to be reported as failed even though that is not actually the case. As a result, the auto-repair scheduler would proceed with scheduling a new repair job and we would end up with a single coordinator node running multiple repair jobs at the same time.

Test Plan: Created a unit test simulating the race condition + deployed the changes to a test instance and verified that auto-repair is still able to run without issue even after a repair failure.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian, dalin, jservin, jchovati

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian, dalin, jservin, jchovati

Subscribers: jchovati, runtian

JIRA Issues: SO-39173

Differential Revision: https://code.uberinternal.com/D16189173