Fix TxnNamedRead discarding paging state
TxnNamedRead was incorrectly determining whether a query was a continuation because the comparison was on TokenKey
and when a MinTokenKey was encountered it would discard the paging state. Comparison on token should be sufficient
because multiple queries in the same batch won't share the same left token bound, only the continuation query will.

Patch by Ariel Weisberg; Reviewed by David Capwell for CASSANDRA-20261