Update RepairTokenRangeSplitter to work with BTI-formatted SSTables
While doing a review pass on the CEP-37 PR I realized that
RepairTokenRangeSplitter could not possibly work with the new BTI
sstable format introduced in 5.0 because it explicitly checks for
BigTableReader implementations.

In addition, the way it calculates the amount of bytes a range covers
in an SSTable is big-format specific.

I noticed that a new utility method
SSTableReader.onDiskSizeForPartitionPositions was added recently in
CASSANDRA-20092 that can effectively calculate this for us in an
implementation-agnostic way.

This change updates the code to use this method, and also remove any
changes we previously made in SSTableScanner and BigTableScanner for
this.

Patch by Andy Tolbert; Reviewed by ___ for CASSANDRA-20315