Use LWTs for all auto_repair_history_v2 mutations
Summary:
We have observed a race condition inside the auto-repair scheduler where auto-repair gets stuck waiting for a replaced node. It seems that there is a conflict between the queries deleting the row about repair data for a replaced node and the query that adds a delete marker to this row. The delete is not executed as an LWT whereas updating the delete markers is.

We believe that this is an issue with how the queries are serialized. The issue can be fixed by manually executing an LWT that deletes the repair history row for the replaced node (non-LWT delete does not work). Therefore we believe that performing all mutations to the `auto_repair_history_v2` should resolve this race condition.

Test Plan:
Deployed the changes to a test instance and verified that the auto-repair scheduler is still able to schedule repair jobs.
{F1037649227}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, panl

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, panl

Subscribers: panl

Tags: #has_java

JIRA Issues: SO-40690

Differential Revision: https://code.uberinternal.com/D16543763