Implement additional metrics on repair failures
Summary: Introduces a set of new repair-related metrics which will make tracking the progress/failures of repair easier. This will be particularly useful for the IR migration, as IR can fail for a number of different reasons each of which require a different follow up. Furthermore, the current implementation of repair requires searching through the logs to determine the reason for failing repair. Having this data as metrics should make the migration process much smoother.

Test Plan:
Unit tests + deployed the changes to a test instance and set up some new Grafana panels:
{F1043717239}

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, jservin, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, jservin, runtian

Subscribers: runtian, dalin, jservin

Tags: #has_java

JIRA Issues: SO-40691

Differential Revision: https://code.uberinternal.com/D16548111