[Cassandra][Code-Yellow] Adds "Driver Name" and "Is Authenticated" to the Emitted Client Metrics
Summary: Adds the tags "driverName" and "IsAutneticated" to the clientsession metric. If a driver name is not provided, the value from the optional field "LANGUAGE" will be used - if it exists.

Test Plan:
Verified by using jconsole on our dev machine:
{F1054158477}

Will be verified through test service after tags are added to the `cassandra-m3-writer`

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, jservin

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, jservin

Subscribers: runtian, jservin

Tags: #has_java

JIRA Issues: SO-41957

Differential Revision: https://code.uberinternal.com/D16903201