Repair turn calculation for bootstrap repair
Summary:
1. For the existing repair types, the repair will be run on multiple nodes on a continuous basis, so we need to calculate precisely
who goes when. However, for the //bootstrap// type, we only need to run repair on the bootstrapping nodes, which will be controlled
by invoking it only during the bootstrap phase. Hence, for the //bootstrap// repair type, it is always its turn when calculated.
2. AutoRepairSchedulerTest: We have to skip the //bootstrap// from InJvm dtest because it short circuits some of the checks. A detailed
InJvm dtest will be added once the full functionality is complete.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, ureview, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Subscribers: runtian

Tags: #has_java

JIRA Issues: SO-43685

Differential Revision: https://code.uberinternal.com/D17363371