A config to run repair for someone's token range
Summary:
So far, AutoRepair has always been assuming the node being repaired is always part of it. So we used to calculate
the token ranges always for the //localhost//. Now, for the bootstrap repair, the node being bootstrapped would
want to run the repair on behalf of the node that just went down.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, ureview, panl, kzalys

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, panl, kzalys

Subscribers: kzalys

Tags: #has_java

JIRA Issues: SO-43685

Differential Revision: https://code.uberinternal.com/D17388629