Add a new bootstrap repair wrapper that manages the e2e flow
Summary:
A new API inside AutoRepairUtilsV2.java, which handles the e2e flow for AutoRepair,
including but not limited to setting up AutoRepair, awaiting for "bootstrap" repair to finish and aborting if
the time interval is more than 15 minutes.

The only thing pending is invoking this newly introduced API from the bootstrap module, which will be done in
subsequent diffs.

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, ureview, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Subscribers: runtian

Tags: #has_java

JIRA Issues: SO-44843

Differential Revision: https://code.uberinternal.com/D17616511