Mutations should not be deferrable on MV lock contention
Summary:
CASSANDRA-12689 mentioned that if a caller is in the mutation stage, it should never defer itself into the same MUTATION queue. If the caller defers the task into the same queue and waits on the future, it’s possible that all available writers are blocked on the future tasks that no one will be able to process.

The introduced change was reverted in following fix stating that throwing WTE for MV mutation will cause repair failure. At Uber we have disabled MV replay on base table repair so this won't be a concern.

Test Plan:
Added distributed test
Test to reproduce the issue without this change: https://code.uberinternal.com/P873879

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, jchovati

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, jchovati

Tags: #has_java

JIRA Issues: SO-45578

Differential Revision: https://code.uberinternal.com/D17918703