CASSANDRA-20709: Get SAI MemtableIndex refs before SSTableIndex refs at query time
SAI predicate search currently has a bug that could result in missing
rows due to a concurrent flush during a query. The new test created in
this PR shows the point of failure. The problem is that we get the
sstable index references before getting the memtable index references.
Note that we do it in the correct order in the ANN OF query path, but
not in the WHERE query path.

This commit updates the QueryView object to hold references to the
appropriate memtable indexes.

It also removes the problematic search methods from MemtableIndexManager
to prevent future misuse.