Introduce AbstractMetricManager and fix AuthMetricsManager
Summary:
Within Uber we created multiple `XMetricManager` classes and they provide very similar functionalities.

Here we introduce an abstract class to standardize the behavior, which requires 2 methods to be implemented:

- `buildKey(Object... parts)` that accept any number of inputs to construct the key instance `K`
- `createMetric(K key)` that return the metric instance `V` from the key instance `K`

A bug in `AuthMetricsManager` is fixed, that the key of the metric map should include the flags as well - otherwise if the enforcement flag / authEnable flag changes after, it won't be reflected in the metric tag and no new metric will be created .

We'll be supporting async metric calls in later diff.

Test Plan: existing UT

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, ureview, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Subscribers: runtian

Tags: #has_java

JIRA Issues: SO-47877

Differential Revision: https://code.uberinternal.com/D18331255